+++ server/src/main/java/org/opensearch/action/ingest/SimulatePipelineTransportAction.java
@@ -38,6 +38,7 @@ import org.opensearch.common.inject.Inject;
 import org.opensearch.common.xcontent.XContentHelper;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.common.io.stream.Writeable;
+import org.opensearch.ingest.IngestPipelineValidator;
 import org.opensearch.ingest.IngestService;
 import org.opensearch.tasks.Task;
 import org.opensearch.threadpool.ThreadPool;
@@ -87,6 +88,8 @@ public class SimulatePipelineTransportAction extends HandledTransportAction<Simu
             listener.onFailure(e);
             return;
         }
+
+        IngestPipelineValidator.validateIngestPipeline(simulateRequest.getPipeline(), ingestService.getClusterService());
 
         executionService.execute(simulateRequest, listener);
     }